track2fasta using cleanpath
